refactor: extract 3 main stage pages and finalize main.tsx shell cleanup (Block F)
